Moved some independent code to modules to reduce clutter.
