+++ Website-Files/modules/logoAnimation.js
@@ -1,0 +1,14 @@
+// This sets up the animation that occurs when the mouse hovers over the logo
+
+export default function logoAnimationSetup(elem){
+    elem.addEventListener('mouseover', () => {
+        elem.style.animation = `1s cubic-bezier(0.77, 0, 0.175, 1) spinLogo`;
+        document.querySelector('.logoText').style.animation = '1s cubic-bezier(0.77, 0, 0.175, 1) slideText';
+        document.querySelector('.logoText').style.transform = 'translateX(0px)';
+    });
+    elem.addEventListener('mouseout', () => {
+        elem.style.animation = `1s cubic-bezier(0.77, 0, 0.175, 1) reverseLogo`;
+        document.querySelector('.logoText').style.animation = '1s cubic-bezier(0.77, 0, 0.175, 1) hideText';
+        document.querySelector('.logoText').style.transform = 'translateX(-250px)';
+    })
+}
