+++ Website-Files/style.css
@@ -661,7 +661,7 @@
     position: fixed;
     color: rgba(255, 255, 255, 1);
     font-size: 10vmin;
-    text-shadow: 0px 0px 10px rgba(255, 255, 255, 0.5);
+    text-shadow: 0px 0px 10px rgba(0, 0, 0, 0.5);
 }
 
 /* ALTITUDE */
